+++ packages/Webkul/Product/src/Repositories/ProductReviewAttachmentRepository.php
@@ -3,11 +3,14 @@
 namespace Webkul\Product\Repositories;
 
 use Webkul\Core\Eloquent\Repository;
+use Webkul\Core\Traits\Sanitizer;
 use Webkul\Product\Contracts\ProductReview;
 use Webkul\Product\Contracts\ProductReviewAttachment;
 
 class ProductReviewAttachmentRepository extends Repository
 {
+    use Sanitizer;
+
     /**
      * Specify model class name.
      */
@@ -22,13 +25,19 @@ class ProductReviewAttachmentRepository extends Repository
     public function upload(array $attachments, ProductReview $review): void
     {
         foreach ($attachments as $attachment) {
-            $fileType = explode('/', $attachment->getMimeType());
+            $mimeType = $attachment->getMimeType();
+
+            $fileType = explode('/', $mimeType);
+
+            $path = $attachment->store('review/'.$review->id);
+
+            $this->sanitizeSVG($path, $mimeType);
 
             $this->create([
-                'path'      => $attachment->store('review/'.$review->id),
+                'path'      => $path,
                 'review_id' => $review->id,
                 'type'      => $fileType[0],
-                'mime_type' => $fileType[1],
+                'mime_type' => $fileType[1] ?? null,
             ]);
         }
     }